ai-buddy/requirements.txt
discord.py
python-dotenv
langchain-groq
langchain-community
duckduckgo-search
ddgs
youtube-transcript-api
beautifulsoup4
requests